api_gateway/requirements.txt
fastapi
httpx
uvicorn
pyjwt
python-multipart
